Changed move time from integer to double to allow increments of less than 1 second

+++ piglet.h
@@ -4,8 +4,8 @@ int devStuff();
 int UCI();
 int perft();
 int perftCheckHash();
-void iterate(std::stop_token st, int depth, int moveTime);
-int calcMoveTime(int moveTime, int wtime, int btime, int winc, int binc, int movestogo);
+void iterate(std::stop_token st, int depth, double moveTime);
+double calcMoveTime(int moveTime, int wtime, int btime, int winc, int binc, int movestogo);
